requirements.txt
streamlit
numpy
onnxruntime
requests
pillow
flask
flask-cors
paho-mqtt